requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
jinja2==3.1.2
